Refactor: move EACRequest model to appropriate directory
